Changed chart parser to return rating counts instead of percentages.

+++ client.go
@@ -163,9 +163,8 @@ func (c *Client) AdvSearch(url string) ([]CollectionItem, error) {
 	return ParseAdvSearch(body)
 }
 
-// RatingPercentages gets the rating percentages (ratings from 1 to 10) for a
-// thing.
-func (c *Client) RatingPercentages(id int) (map[int]float64, error) {
+// RatingPercentages gets the rating counts (ratings from 1 to 10) for a thing.
+func (c *Client) RatingPercentages(id int) (map[int]int, error) {
 	resp, err := c.httpClient.Get(fmt.Sprintf(
 		"http://boardgamegeek.com/graphstats/thing/%d?ajax=1",
 		id,
